rename folder util to pages and adding scroll at statics_page
